separate file for setters
