+++ farm-staking/farm-staking-nft/wasm/src/lib.rs
@@ -23,12 +23,6 @@ multiversx_sc_wasm_adapter::endpoints! {
         upgrade => upgrade
         mergeFarmTokens => merge_farm_tokens_endpoint
         calculateRewardsForGivenPosition => calculate_rewards_for_given_position
-        topUpRewards => top_up_rewards
-        withdrawRewards => withdraw_rewards
-        endProduceRewards => end_produce_rewards
-        setPerBlockRewardAmount => set_per_block_rewards
-        setMaxApr => set_max_apr
-        setMinUnbondEpochs => set_min_unbond_epochs_endpoint
         startProduceRewards => start_produce_rewards_endpoint
         getAccumulatedRewards => accumulated_rewards
         getRewardCapacity => reward_capacity
@@ -91,6 +85,12 @@ multiversx_sc_wasm_adapter::endpoints! {
         removeHook => remove_hook
         registerUnbondToken => register_unbond_token
         getUnbondTokenId => unbond_token
+        topUpRewards => top_up_rewards
+        withdrawRewards => withdraw_rewards
+        endProduceRewards => end_produce_rewards
+        setPerBlockRewardAmount => set_per_block_rewards
+        setMaxApr => set_max_apr
+        setMinUnbondEpochs => set_min_unbond_epochs_endpoint
     )
 }
